docs(usb-keyboard-example): clean up the USB keyboard example
